+++ shuup/order_printouts/admin_module/__init__.py
@@ -20,8 +20,18 @@ class PrintoutsAdminModule(AdminModule):
                 name="printouts.delivery_pdf"
             ),
             admin_url(
+                "^printouts/delivery/email/(?P<shipment_pk>\d+)/$",
+                "shuup.order_printouts.admin_module.views.send_delivery_email",
+                name="printouts.delivery_email"
+            ),
+            admin_url(
                 "^printouts/confirmation/(?P<order_pk>\d+)/$",
                 "shuup.order_printouts.admin_module.views.get_confirmation_pdf",
                 name="printouts.confirmation_pdf"
+            ),
+            admin_url(
+                "^printouts/confirmation/email/(?P<order_pk>\d+)/$",
+                "shuup.order_printouts.admin_module.views.send_confirmation_email",
+                name="printouts.confirmation_email"
             ),
         ]