+++ shuup/order_printouts/admin_module/section.py
@@ -9,12 +9,21 @@ from django.utils.translation import ugettext_lazy as _
 
 from shuup.admin.base import Section
 
+from .forms import PrintoutsEmailForm
+
+
+EMAIL_DEFAULT_BODY = _("""Important information regarding your order, see attachment.
+
+Best Regards,
+%(shop)s""")
+
 
 class PrintoutsSection(Section):
     identifier = "printouts_section"
     name = _("Printouts")
     icon = "fa-print"
     template = "shuup/order_printouts/admin/section.jinja"
+    extra_js = "shuup/order_printouts/admin/section_js.jinja"
     order = 5
 
     @staticmethod
@@ -23,4 +32,14 @@ class PrintoutsSection(Section):
 
     @staticmethod
     def get_context_data(obj):
-        return {}
+        recipient = None
+        if obj.customer:
+            recipient = obj.customer.email
+        elif obj.billing_address:
+            recipient = obj.billing_address.email
+        data = {
+            "to": recipient,
+            "subject": _("%(shop)s: Order %(pk)s") % {"shop": obj.shop.name, "pk": obj.pk},
+            "body": (EMAIL_DEFAULT_BODY % {"shop": obj.shop.name}).strip()
+        }
+        return {"email_form": PrintoutsEmailForm(initial=data)}