+++ doc/changelog.rst
@@ -36,6 +36,7 @@ Simple Supplier
 Order Printouts
 ~~~~~~~~~~~~~~~
 
+- Add options to send printouts as email attachments
 - Move printouts to tab from toolbar
 
 Campaigns